integrated model classes from Alexandre
