+++ app/src/main/java/ch/epfl/skysync/dataModels/calendarModels/CalendarViewable.kt
@@ -1,0 +1,8 @@
+package ch.epfl.skysync.dataModels.calendarModels
+
+import java.time.LocalDate
+
+interface CalendarViewable {
+    val date: LocalDate
+    val timeSlot: TimeSlot
+}
