+++ app/src/main/java/ch/epfl/skysync/dataModels/calendarModels/TimeSlot.kt
@@ -1,5 +1,7 @@
 package ch.epfl.skysync.dataModels.calendarModels
 
+/** Represent the day in two time slots, morning (AM) and afternoon (PM) */
 enum class TimeSlot {
-    AM, PM
+  AM,
+  PM
 }
