+++ app/src/main/java/ch/epfl/skysync/database/schemas/AvailabilitySchema.kt
@@ -1,9 +1,9 @@
 package ch.epfl.skysync.database.schemas
 
 import ch.epfl.skysync.database.Schema
-import ch.epfl.skysync.model.Availability
-import ch.epfl.skysync.model.AvailabilityStatus
-import ch.epfl.skysync.model.TimeSlot
+import ch.epfl.skysync.dataModels.calendarModels.Availability
+import ch.epfl.skysync.dataModels.calendarModels.AvailabilityStatus
+import ch.epfl.skysync.dataModels.calendarModels.TimeSlot
 import com.google.firebase.firestore.DocumentId
 import java.time.ZoneOffset
 import java.util.Date
