--- app/src/main/java/ch/epfl/skysync/model/TimeSlot.kt
@@ -1,7 +1,0 @@
-package ch.epfl.skysync.model
-
-/** Represent the day in two time slots, morning (AM) and afternoon (PM) */
-enum class TimeSlot {
-  AM,
-  PM
-}
