+++ app/src/test/java/ch/epfl/skysync/database/tables/AvailabilityTableUnitTest.kt
@@ -2,9 +2,9 @@ package ch.epfl.skysync.database.tables
 
 import ch.epfl.skysync.database.MockDatabase
 import ch.epfl.skysync.database.schemas.AvailabilitySchema
-import ch.epfl.skysync.model.Availability
-import ch.epfl.skysync.model.AvailabilityStatus
-import ch.epfl.skysync.model.TimeSlot
+import ch.epfl.skysync.dataModels.calendarModels.Availability
+import ch.epfl.skysync.dataModels.calendarModels.AvailabilityStatus
+import ch.epfl.skysync.dataModels.calendarModels.TimeSlot
 import java.time.LocalDate
 import org.junit.Assert.*
 import org.junit.Test
